requirements.txt
apt-clone==0.2.1
apt-xapian-index==0.47
apturl==0.5.2
blinker==1.3
Brlapi==0.6.4
bz2file==0.98
CacheControl==0.11.5
chardet==2.3.0
command-not-found==0.3
config==0.4.0
contextlib2==0.5.1
cryptography==1.2.3
cycler==0.9.0
decorator==4.0.6
defer==1.0.6
dirspec==13.10
future==0.15.2
httplib2==0.9.1
idna==2.0
ipython==2.4.1
joblib==0.9.4
lockfile==0.12.2
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
MarkupSafe==0.23
matplotlib==1.5.1
natsort==4.0.3
networkx==1.11
nose==1.3.7
numpy==1.11.0
oauthlib==1.0.3
onboard==1.2.0
oneconf==0.3.9
pandas==0.17.1
pbr==5.1.3
pexpect==4.0.1
Pillow==3.1.2
piston-mini-client==0.7.5
pma-python==2.0.0.60
psutil==3.4.2
ptyprocess==0.5
pyasn1==0.1.9
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0
pygobject==3.20.0
pygraphviz==1.3.1
PyICU==1.9.2
PyJWT==1.3.0
pyparsing==2.0.3
python-apt==1.1.0b1+ubuntu0.16.4.5
python-dateutil==2.4.2
python-debian==0.1.27
python-igraph==0.7.1.post6
python-xapp==1.0.0
pytz==2014.10
pyxdg==0.25
reportlab==3.3.0
requests==2.9.1
scikit-bio==0.4.1
scikit-image==0.10.1
scikit-learn==0.17
scipy==0.17.0
screen-resolution-extra==0.0.0
sessioninstaller==0.0.0
setproctitle==1.1.8
simplegeneric==0.8.1
six==1.10.0
sklearn==0.0
stevedore==1.30.1
system-service==0.3
thin-client-config-agent==0.8
ubuntu-drivers-common==0.0.0
ufw==0.35
urllib3==1.13.1
virtkey==0.63.0
virtualenv==16.4.3
virtualenv-clone==0.5.2
virtualenvwrapper==4.8.4
xkit==0.0.0
